% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractBufferValues.R
\name{aggregateBufferValues}
\alias{aggregateBufferValues}
\title{Aggregate buffer values}
\usage{
aggregateBufferValues(values, varName, fun)
}
\arguments{
\item{values}{Values extracted from buffers, including an "ID" column}

\item{varName}{Name pattern of the variable to aggregate by}

\item{fun}{A function to aggregate buffer values by: max, min}
}
\value{
A dataframe of buffer values.
}
\description{
Groups values by buffer and--for each buffer--keeps the entry
with the "fun" (min/max) variable value.
}
