% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAnalysisRegionMask.R
\name{createAnalysisRegionMask}
\alias{createAnalysisRegionMask}
\title{Create analysis region mask}
\usage{
createAnalysisRegionMask(raster, initRange)
}
\arguments{
\item{raster}{A SpatRaster of explanatory variables}

\item{initRange}{A matrix that holds the min & max initiation limits of each
explanatory variable}
}
\value{
A SpatRaster with NA cells anywhere outside the analysis region.
}
\description{
Creates a raster mask that identifies area(s) with landslide
initiation characteristics. Cells with variable values outside their
respective initiation ranges are marked with NA.
}
