% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAverageRaster.R
\name{createAverageRaster}
\alias{createAverageRaster}
\title{Create average raster}
\usage{
createAverageRaster(rasterFiles)
}
\arguments{
\item{rasterFiles}{a vector of raster files to average together}
}
\value{
a raster object.
}
\description{
Creates an average raster from a set of raster files. Good for
averaging many large raster files that might not fit into memory all at once.
}
