% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createInitiationRange.R
\name{createInitiationRange}
\alias{createInitiationRange}
\title{Create initiation range}
\usage{
createInitiationRange(raster, initBuffers, expansion)
}
\arguments{
\item{raster}{A SpatRaster of explanatory variables}

\item{initBuffers}{A SpatVector of initiation buffers}

\item{expansion}{Percent to reduce each range min and increase each range
max by}
}
\value{
A matrix that holds the min & max initiation limits of each raster
variable.
}
\description{
Finds the initiation range for each raster variable.
}
