% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createNoninitiationRaster.R
\name{createNoninitiationRaster}
\alias{createNoninitiationRaster}
\title{Create non-initiation raster}
\usage{
createNoninitiationRaster(raster, initRange, initBuffers)
}
\arguments{
\item{raster}{A SpatRaster of explanatory variables}

\item{initRange}{A matrix that holds the min & max initiation limits
of each explanatory variable}

\item{initBuffers}{A SpatVector of initiation buffers}
}
\value{
A SpatRaster where any non-NA cell is a viable location for a
non-initiation point.
}
\description{
Creates a raster that defines where non-initiation points can be
generated. This includes cells not already in initiation sites and whose
variable values are all within their respective initiation ranges.
}
