% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPropRaster.R
\name{createPropRaster}
\alias{createPropRaster}
\title{create proportion raster}
\usage{
createPropRaster(probRaster)
}
\arguments{
\item{probRaster}{probability raster}
}
\value{
A proportion raster
}
\description{
Creates a proportion raster from a probability raster
}
