% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractBufferValues.R
\name{extractBufferValues}
\alias{extractBufferValues}
\title{Extract buffer values}
\usage{
extractBufferValues(raster, initBuffers, noninitBuffers, extractionMethod)
}
\arguments{
\item{raster}{A SpatRaster of explanatory variables}

\item{initBuffers}{A SpatVector of initiation buffers}

\item{noninitBuffers}{A SpatVector of non-initiation buffers}

\item{extractionMethod}{Method to use for extracting values from each buffer:
"all cells", "center cell", "max gradient cell", or
"max plan cell"}
}
\value{
A dataframe of extracted raster values with an additional "class"
column.
}
\description{
Extracts variable raster values within initiation and
non-initiation buffers.
}
