% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateNoninitiationBuffers.R
\name{generateNoninitiationBuffers}
\alias{generateNoninitiationBuffers}
\title{Generate non-initiation buffers}
\usage{
generateNoninitiationBuffers(count, region, radius)
}
\arguments{
\item{count}{The number of non-initiation buffers to generate}

\item{region}{A SpatRaster with NA cells anywhere a non-initiation buffer
cannot be located}

\item{radius}{The radius of each non-initiation buffer}
}
\value{
A SpatVector of non-initiation buffers.
}
\description{
Generates a SpatVector of non-initiation buffers in a given
area.
}
