% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeDataFrame.R
\name{summarizeDataFrame}
\alias{summarizeDataFrame}
\title{Summarize a data frame}
\usage{
summarizeDataFrame(x)
}
\arguments{
\item{x}{the \code{data.frame} to summarize}
}
\value{
a \code{matrix} of summary statistics
}
\description{
Summarize a data frame
}
\details{
Calculates summary statistics for each data frame column and stores
them in a formatted matrix.
}
