% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeVector.R
\name{summarizeVector}
\alias{summarizeVector}
\title{Summarize a numeric vector}
\usage{
summarizeVector(x)
}
\arguments{
\item{x}{the numeric vector to summarize}
}
\value{
a \code{matrix} of summary statistics
}
\description{
Summarize a numeric vector
}
\details{
Calculates summary statistics for a numeric vector and stores
them in a formatted matrix.
}
